Bug 3587: removing a sub-task from an Expansion does not in itself force the Expansion to re-calculate the Reported result, even though the AllocationResultAggregator will often compute a different value. Usually one of the other sub-tasks will change for some other reason, forcing a Notification to be sent, and therefore recalculating the result. But that may never happen.
This fix creates a new helper method in Plugin Helper.
Plugins that want to remove a task from an expansion should
1) Identify the task to remove
2) Call this helper method
3) publishRemove the task (or re-parent it if that is the desired
effect)
---- Note particularly that it remains the Plugins job to remove the
task. This method is solely responsible for ensuring that the
Expansion's Reported AllocationResult is updated if necessary, and the
Expansion is publishChanged if necessary that AR has changed.