Bug 3657: ExpansionImpl: When auto-propogating removal of Expansions, clear the back-pointers to the parent task (which were already being cleared when _not_ auto-propogating.
Additionally, clear the task.getWorkflow() pointer on the sub-tasks
which will be publishRemoved during auto-propogation.

Finally, add a comment under the auto-propogation, indicating the
desirability of clearing the task.getWorkflow pointer.

TaskImpl:
When publishRemoving a Task, if it still has a pointer to a Workflow
(regardless of whether it was still in the workflow), then clear that
back-pointer.

The net result is that all the various pointers should get cleared up
on publishRemove - when auto-propogating, we temporarily leave the
task.getWorkflow back-pointer in case any plugins would like to use it.

Note that the clearing of task.getWorkflow means that plugins handling
these sub-tasks must not assume that a task that was once (ie when the
execute cycle started) a member of an Expansion, will still have a
task.getWorkflow reference, by the time they look at it.